01_docker_compose/app/requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.20.0
django-split-settings==1.1.0
asgiref==3.5.0
attrs==21.4.0
django-debug-toolbar==3.3.0
Faker==13.4.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uWSGI==2.0.20
gunicorn==20.1.0